RE-345 Split Up BetterBags Events to its own events
