+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/events/bag/EntityPickupItemBags.java
@@ -1,0 +1,73 @@
+package de.relluem94.minecraft.server.spigot.essentials.events.bag;
+
+import de.relluem94.minecraft.server.spigot.essentials.CustomItems;
+import de.relluem94.minecraft.server.spigot.essentials.RelluEssentials;
+import de.relluem94.minecraft.server.spigot.essentials.constants.MessageKey;
+import de.relluem94.minecraft.server.spigot.essentials.constants.NamespacedKeyConstants;
+import de.relluem94.minecraft.server.spigot.essentials.helpers.BagHelper;
+import de.relluem94.minecraft.server.spigot.essentials.helpers.ChatHelper;
+import de.relluem94.minecraft.server.spigot.essentials.helpers.StringHelper;
+import de.relluem94.minecraft.server.spigot.essentials.helpers.pojo.PlayerEntry;
+import org.bukkit.Sound;
+import org.bukkit.SoundCategory;
+import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.Listener;
+import org.bukkit.event.entity.EntityPickupItemEvent;
+import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.ItemMeta;
+import org.bukkit.persistence.PersistentDataType;
+import org.jetbrains.annotations.NotNull;
+
+import static de.relluem94.minecraft.server.spigot.essentials.RelluEssentials.languageHelper;
+import static de.relluem94.minecraft.server.spigot.essentials.constants.NamespacedKeyConstants.itemCoins;
+
+public class EntityPickupItemBags implements Listener {
+
+    @EventHandler
+    public void onItemCollect(@NotNull EntityPickupItemEvent e) {
+        if (e.getEntity() instanceof Player p) {
+
+            PlayerEntry pe = RelluEssentials.getInstance().getPlayerAPI().getPlayerEntry(p);
+
+            ItemStack is = e.getItem().getItemStack();
+            if (CustomItems.coins.almostEquals(is)) {
+                ItemMeta im = is.getItemMeta();
+
+                if (im != null && im.getPersistentDataContainer().has(itemCoins(), PersistentDataType.INTEGER)) {
+                    Integer itemCoins = im.getPersistentDataContainer().get(NamespacedKeyConstants.itemCoins(), PersistentDataType.INTEGER);
+
+                    if (itemCoins == null) {
+                        itemCoins = 0;
+                    }
+
+                    int coins = itemCoins * is.getAmount();
+                    ChatHelper.sendMessageInActionBar(p, languageHelper.getWithPrefix(MessageKey.COMMAND_PURSE_GAIN, StringHelper.formatInt(coins), StringHelper.formatDouble(pe.getPurse() + coins)));
+                    pe.setPurse(pe.getPurse() + coins);
+
+                    p.playSound(p, Sound.ITEM_ARMOR_EQUIP_GOLD, SoundCategory.PLAYERS, 1F, 1);
+
+                    pe.setUpdatedBy(pe.getId());
+                    pe.setHasToBeUpdated(true);
+
+                    e.getItem().getItemStack().setAmount(0);
+                    e.setCancelled(true);
+                }
+            }
+
+            String worldName = p.getWorld().getName();
+            boolean collectBagEnabled = RelluEssentials.getInstance().collectBagWorlds.contains(worldName);
+
+            if (collectBagEnabled && BagHelper.hasBags(pe.getId()) && BagHelper.collectItem(e.getItem(), p, pe)) {
+                p.getInventory().remove(is);
+                e.setCancelled(true);
+                e.getItem().remove();
+            } else {
+                p.getInventory().addItem(is);
+                e.setCancelled(true);
+                e.getItem().remove();
+                p.playSound(p, Sound.ENTITY_ITEM_PICKUP, SoundCategory.PLAYERS, 0.5f, 1f);
+            }
+        }
+    }
+}
