+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/managers/EventManager.java
@@ -3,6 +3,10 @@ package de.relluem94.minecraft.server.spigot.essentials.managers;
 import de.relluem94.minecraft.server.spigot.essentials.RelluEssentials;
 import de.relluem94.minecraft.server.spigot.essentials.constants.MessageKey;
 import de.relluem94.minecraft.server.spigot.essentials.events.*;
+import de.relluem94.minecraft.server.spigot.essentials.events.bag.BlockBreakBags;
+import de.relluem94.minecraft.server.spigot.essentials.events.bag.BlockDropItemBags;
+import de.relluem94.minecraft.server.spigot.essentials.events.bag.EntityPickupItemBags;
+import de.relluem94.minecraft.server.spigot.essentials.events.bag.InventoryClickBags;
 import de.relluem94.minecraft.server.spigot.essentials.events.npc.DamgeNPC;
 import de.relluem94.minecraft.server.spigot.essentials.events.npc.InteractNPC;
 import de.relluem94.minecraft.server.spigot.essentials.events.npc.InventoryClickNPC;
@@ -28,7 +32,10 @@ public class EventManager implements IEnable{
         pm.registerEvents(new BetterPlayerQuit(), RelluEssentials.getInstance());           eventCount++;
         pm.registerEvents(new BetterBlockDrop(), RelluEssentials.getInstance());            eventCount++;
         pm.registerEvents(new BetterLights(), RelluEssentials.getInstance());               eventCount++;
-        pm.registerEvents(new BetterBags(), RelluEssentials.getInstance());                 eventCount++;
+        pm.registerEvents(new BlockBreakBags(), RelluEssentials.getInstance());             eventCount++;
+        pm.registerEvents(new BlockDropItemBags(), RelluEssentials.getInstance());          eventCount++;
+        pm.registerEvents(new InventoryClickBags(), RelluEssentials.getInstance());         eventCount++;
+        pm.registerEvents(new EntityPickupItemBags(), RelluEssentials.getInstance());       eventCount++;
         pm.registerEvents(new BlockPlace(), RelluEssentials.getInstance());                 eventCount++;
         pm.registerEvents(new BetterMobs(), RelluEssentials.getInstance());                 eventCount++;
         pm.registerEvents(new BetterSoil(), RelluEssentials.getInstance());                 eventCount++;
